requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
python-dotenv>=0.5.1
absl-py==1.4.0
aiohttp 
aiosignal 
alembic==1.9.4
anyio 
argon2-cffi 
argon2-cffi-bindings 
arrow==1.2.3
asttokens 
async-timeout 
attrs 
Babel==2.11.0
backcall 
backports.functools-lru-cache 
beautifulsoup4 
black==22.12.0
bleach 
brotlipy 
cachetools==5.3.0
certifi==2022.12.7
cffi 
charset-normalizer 
click==8.1.3
cloudpickle==2.2.1
colorama 
contourpy 
cryptography 
cycler 
databricks-cli==0.17.4
debugpy 
decorator 
defusedxml 
docker==6.0.1
entrypoints 
executing 
fastjsonschema 
Flask==2.2.3
flit_core 
fonttools 
fqdn==1.5.1
frozenlist 
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.0
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.1
idna 
imbalanced-learn 
importlib-metadata 
importlib-resources 
ipykernel 
ipython 
ipython-genutils==0.2.0
ipywidgets 
isoduration==20.11.0
isort==5.11.4
itsdangerous==2.1.2
jedi 
Jinja2 
joblib 
json5==0.9.11
jsonpointer==2.3
jsonschema 
jupyter 
jupyter-console 
jupyter-events 
jupyter_client 
jupyter_core 
jupyter_server 
jupyter_server_terminals 
jupyterlab==3.5.2
jupyterlab-code-formatter==1.5.3
jupyterlab-pygments 
jupyterlab-spellchecker==0.7.2
jupyterlab-widgets 
jupyterlab_server==2.18.0
kiwisolver 
llvmlite==0.39.1
loguru==0.6.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe 
matplotlib 
matplotlib-inline 
mistune 
mlflow==2.1.1
multidict 
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic 
nbclient 
nbconvert 
nbformat 
nest-asyncio 
notebook 
notebook_shim 
numba==0.56.4
numpy 
oauthlib==3.2.2
packaging 
pandas 
pandocfilters 
parso 
pathspec==0.10.3
patsy 
pickleshare 
Pillow 
pkgutil_resolve_name 
platformdirs 
ply==3.11
prometheus-client 
prompt-toolkit 
protobuf==3.20.3
psutil 
pure-eval 
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser 
Pygments 
PyJWT==2.6.0
pyOpenSSL 
pyparsing 
PyQt5==5.15.7
PyQt5-sip 
pyrsistent 
PySocks 
python-dateutil 
python-json-logger 
pytz 
pywin32==304
pywinpty 
PyYAML 
pyzmq 
qtconsole 
QtPy 
querystring-parser==1.2.4
requests 
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn>=1.2.2
scipy==1.9.3
seaborn 
Send2Trash 
shap==0.41.0
sip 
six 
slicer==0.0.7
smmap==5.0.0
sniffio 
soupsieve 
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data 
statsmodels 
tabulate 
terminado 
threadpoolctl 
tinycss2 
toml 
tomli==2.0.1
torch==2.0.0
torchvision==0.14.0
tornado 
tqdm 
traitlets 
typing_extensions 
unicodedata2 
uri-template==1.2.0
urllib3 
waitress==2.1.2
wcwidth 
webcolors==1.12
webencodings==0.5.1
websocket-client 
Werkzeug==2.2.2
widgetsnbextension 
win-inet-pton 
win32-setctime==1.1.0
yarl 
zipp 

